make objects serializable without running through YamlSerializer
