+++ xnmt/batcher.py
@@ -3,7 +3,7 @@ import random
 import numpy as np
 import dynet as dy
 from xnmt.vocab import Vocab
-from xnmt.serialize.serializable import Serializable
+from xnmt.serialize.serializable import Serializable, serializable_init
 
 class Batch(list):
   """
@@ -136,6 +136,8 @@ class InOrderBatcher(Batcher, Serializable):
     pad_src_to_multiple (int): pad the source side so that the batch length is a multiple of the specified integer
   """
   yaml_tag = "!InOrderBatcher"
+
+  @serializable_init
   def __init__(self, batch_size, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                pad_src_to_multiple=1):
     super(InOrderBatcher, self).__init__(batch_size, src_pad_token=src_pad_token,
@@ -224,6 +226,8 @@ class SrcBatcher(SortBatcher, Serializable):
     pad_src_to_multiple (int): pad the source side so that the batch length is a multiple of the specified integer
   """
   yaml_tag = "!SrcBatcher"
+
+  @serializable_init
   def __init__(self, batch_size, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                break_ties_randomly=True, pad_src_to_multiple=1):
     super(SrcBatcher, self).__init__(batch_size, sort_key=lambda x: len(x[0]), granularity='sent',
@@ -243,6 +247,8 @@ class TrgBatcher(SortBatcher, Serializable):
     pad_src_to_multiple (int): pad the source side so that the batch length is a multiple of the specified integer
   """
   yaml_tag = "!TrgBatcher"
+
+  @serializable_init
   def __init__(self, batch_size, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                break_ties_randomly=True, pad_src_to_multiple=1):
     super(TrgBatcher, self).__init__(batch_size, sort_key=lambda x: len(x[1]), granularity='sent',
@@ -262,6 +268,8 @@ class SrcTrgBatcher(SortBatcher, Serializable):
     pad_src_to_multiple (int): pad the source side so that the batch length is a multiple of the specified integer
   """
   yaml_tag = "!SrcTrgBatcher"
+
+  @serializable_init
   def __init__(self, batch_size, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                break_ties_randomly=True, pad_src_to_multiple=1):
     super(SrcTrgBatcher, self).__init__(batch_size, sort_key=lambda x: len(x[0])+1.0e-6*len(x[1]),
@@ -282,6 +290,8 @@ class TrgSrcBatcher(SortBatcher, Serializable):
     pad_src_to_multiple (int): pad the source side so that the batch length is a multiple of the specified integer
   """
   yaml_tag = "!TrgSrcBatcher"
+
+  @serializable_init
   def __init__(self, batch_size, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                break_ties_randomly=True, pad_src_to_multiple=1):
     super(TrgSrcBatcher, self).__init__(batch_size, sort_key=lambda x: len(x[1])+1.0e-6*len(x[0]),
@@ -302,6 +312,7 @@ class SentShuffleBatcher(ShuffleBatcher, Serializable):
   """
   yaml_tag = "!SentShuffleBatcher"
 
+  @serializable_init
   def __init__(self, batch_size, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                pad_src_to_multiple=1):
     super(SentShuffleBatcher, self).__init__(batch_size, granularity='sent', src_pad_token=src_pad_token,
@@ -320,12 +331,13 @@ class WordShuffleBatcher(ShuffleBatcher, Serializable):
   """
   yaml_tag = "!WordShuffleBatcher"
 
+  @serializable_init
   def __init__(self, words_per_batch, src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES,
                pad_src_to_multiple=1):
     super(WordShuffleBatcher, self).__init__(words_per_batch, granularity='word', src_pad_token=src_pad_token,
                                              trg_pad_token=trg_pad_token, pad_src_to_multiple=pad_src_to_multiple)
 
-class WordSortBatcher(SortBatcher, Serializable):
+class WordSortBatcher(SortBatcher):
   """
   Base class for word sort-based batchers
   """
@@ -358,6 +370,7 @@ class WordSrcBatcher(WordSortBatcher, Serializable):
   """
   yaml_tag = "!WordSrcBatcher"
 
+  @serializable_init
   def __init__(self, words_per_batch=None, avg_batch_size=None,
                src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES, break_ties_randomly=True,
                pad_src_to_multiple=1):
@@ -385,6 +398,7 @@ class WordTrgBatcher(WordSortBatcher, Serializable):
   """
   yaml_tag = "!WordTrgBatcher"
 
+  @serializable_init
   def __init__(self, words_per_batch=None, avg_batch_size=None,
                src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES, break_ties_randomly=True,
                pad_src_to_multiple=1):
@@ -412,6 +426,7 @@ class WordSrcTrgBatcher(WordSortBatcher, Serializable):
   """
   yaml_tag = "!WordSrcTrgBatcher"
 
+  @serializable_init
   def __init__(self, words_per_batch=None, avg_batch_size=None,
                src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES, break_ties_randomly=True,
                pad_src_to_multiple=1):
@@ -439,6 +454,7 @@ class WordTrgSrcBatcher(WordSortBatcher, Serializable):
   """
   yaml_tag = "!WordTrgSrcBatcher"
 
+  @serializable_init
   def __init__(self, words_per_batch=None, avg_batch_size=None,
                src_pad_token=Vocab.ES, trg_pad_token=Vocab.ES, break_ties_randomly=True,
                pad_src_to_multiple=1):
