+++ xnmt/hyper_parameters.py
@@ -1,12 +1,13 @@
 
 from xnmt.events import register_xnmt_handler, handle_xnmt_event
 from xnmt.reports import Reportable
-from xnmt.serialize.serializable import Serializable
+from xnmt.serialize.serializable import Serializable, serializable_init
 
 class ScalingParam(Serializable):
   ''' initial * scaler(epoch-1) '''
   yaml_tag = "!ScalingParam"
 
+  @serializable_init
   def __init__(self, initial=0.0, scaler=None):
     self.__value = initial
     self.scaler = scaler
@@ -23,6 +24,7 @@ class ScalingParam(Serializable):
 class Scalar(Serializable):
   yaml_tag = "!Scalar"
 
+  @serializable_init
   def __init__(self, initial=0.0):
     self.__value = initial
 
@@ -38,6 +40,7 @@ class GeometricSequence(Serializable):
 
   # Do not set warmup_counter manually.
   @register_xnmt_handler
+  @serializable_init
   def __init__(self, initial=0.1, warmup=0, ratio=1, min_value=0.0, max_value=1.0):
     self.__value = initial
     self.warmup = warmup
@@ -67,6 +70,7 @@ class GeometricSequence(Serializable):
 class DefinedSequence(Serializable):
   yaml_tag = '!DefinedSequence'
   @register_xnmt_handler
+  @serializable_init
   def __init__(self, sequence=None):
     assert sequence is not None
     assert type(sequence) == list, "DefinedSequence need to have a list type"
