+++ xnmt/param_init.py
@@ -3,7 +3,7 @@ import math
 import numpy as np
 import dynet as dy
 
-from xnmt.serialize.serializable import Serializable
+from xnmt.serialize.serializable import Serializable, serializable_init
 
 class ParamInitializer(object):
   """
@@ -34,6 +34,8 @@ class NormalInitializer(ParamInitializer, Serializable):
     var (float): Variance of the distribution
   """
   yaml_tag = "!NormalInitializer"
+
+  @serializable_init
   def __init__(self, mean=0, var=1):
     self.mean = mean
     self.var = var
@@ -49,6 +51,8 @@ class UniformInitializer(ParamInitializer, Serializable):
     scale (float): Parameters are sampled from :math:`\mathcal U([-\\texttt{scale},\\texttt{scale}])`
   """
   yaml_tag = "!UniformInitializer"
+
+  @serializable_init
   def __init__(self, scale):
     self.scale = scale
   def initializer(self, dim, is_lookup=False, num_shared=1):
@@ -64,6 +68,8 @@ class ConstInitializer(ParamInitializer, Serializable):
     c (float): Value to initialize the parameters
   """
   yaml_tag = "!ConstInitializer"
+
+  @serializable_init
   def __init__(self, c):
     self.c = c
   def initializer(self, dim, is_lookup=False, num_shared=1):
@@ -92,6 +98,8 @@ class GlorotInitializer(ParamInitializer, Serializable):
     gain (float): Gain (Depends on the activation function)
   """
   yaml_tag = "!GlorotInitializer"
+
+  @serializable_init
   def __init__(self, gain=1.0):
     self.gain = gain
   def initializer(self, dim, is_lookup=False, num_shared=1):
@@ -124,6 +132,8 @@ class FromFileInitializer(ParamInitializer, Serializable):
     fname (str): File name
   """
   yaml_tag = "!FromFileInitializer"
+
+  @serializable_init
   def __init__(self, fname):
     self.fname = fname
   def initializer(self, dim, is_lookup=False, num_shared=1):
@@ -141,6 +151,8 @@ class NumpyInitializer(ParamInitializer, Serializable):
     array (np.ndarray): Numpy array
   """
   yaml_tag = "!NumpyInitializer"
+
+  @serializable_init
   def __init__(self, array):
     self.array = array
   def initializer(self, dim, is_lookup=False, num_shared=1):
@@ -154,6 +166,11 @@ class ZeroInitializer(ParamInitializer, Serializable):
   Initializes parameter matrix to zero (most appropriate for bias parameters).
   """
   yaml_tag="!ZeroInitializer"
+
+  @serializable_init
+  def __init__(self):
+    pass
+
   def initializer(self, dim, is_lookup=False, num_shared=1):
     return dy.ConstInitializer(c=0.0)
 
@@ -166,8 +183,11 @@ class LeCunUniformInitializer(ParamInitializer, Serializable):
     scale (float): scale
   """
   yaml_tag = "!LeCunUniformInitializer"
+
+  @serializable_init
   def __init__(self, scale=1.0):
     self.scale = scale
+
   def initializer(self, dim, is_lookup=False, num_shared=1):
     if is_lookup:
       fan_in = dim[0]
